scripts/auto_tag/pyproject.toml
[tool.poetry]
name = "auto_tag"
version = "0.1.0"
description = ""
authors = ["Tom Dugan <tom.dugan@affirm.com>"]

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "^3.1.27"
PyYAML = "^6.0"
regex = "^2022.8.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
